requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
asyncpg==0.31.0
babel==2.18.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
courlan==1.3.2
cryptography==46.0.6
dateparser==1.4.0
distro==1.9.0
fastapi==0.135.2
feedparser==6.0.12
google-auth==2.49.1
google-genai==1.68.0
greenlet==3.3.2
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
regex==2026.3.32
requests==2.33.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
alembic==1.14.0
starlette==1.0.0
tenacity==9.1.4
tld==0.13.2
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
